[BUGFIX] Doctrine@PostLoad calls DI and AOP initialization explicitly.
When a Doctrine lazy loading proxy for an object is
created for some specific purpose (e.g. when initializing
session objects relating to an entity) but filled afterwards
"on the fly" if this object is part of a wide range "findBy"
call through some query, Doctrine initializes the lazy
loading proxy but does not call __wakeup and therefore not initialize AOP and DI.
We work around this issue, by initializing those objects
with a Doctrine PostLoad hook explicitly.

Change-Id: I83ba6484e9daecf5344242815ec52ad1770907f9
Resolves: #30467
Releases: 1.1