[FEATURE] Allow translation files (XLIFF) on a per package basis
Fixes issue in which XliffTranslationProvider prefaces a hardcoded
base path of resource://FLOW3/Private/Locale/Translations before
the sourceName argument passed to getTranslationByOriginalLabel and
getTranslationById methods.

This fix modifies the translateById and translateByOriginalLabel
methods in the TYPO3\FLOW3\I18n\Translator to use packageKey
as an argument. The XliffTranslationProvider will construct a
base path of "<packageName>/Private/Locale/Translations". This base
path is prepended to the value of the sourceName argument to create
the translation file resource string.

Unit tests have been adjusted as well to accommodate these changes
to the XliffTranslationProvider and Translator classes. The
TranslationProviderInterface has been adjusted as well.

Change-Id: Iff62c0dc4cf47f8b2d5a9d143132ab64e2012891
Resolves: #14037
Releases: 1.1