[FEATURE] Support mapping to specific subtypes by property (__type)
When type hierarchies are used for modeling, it's sometimes necessary
to use a special subclass when creating new entities (e.g. from user
selection). This change enables the use of a special (request)
property "__type" in the source array of a type converter. For security
reasons, this feature must be configured explicitly through a mapping
configuration.

Resolves: #30300
Releases: 1.1
Change-Id: I6d0a9d62789bd8557a848311e1f13f0c094dc51f