[FEATURE][!!!] Modular and configurable bootstrap
This extracts all specific initialization methods from the bootstrap
into a "Scripts" class and instead of hard-wiring each initialization
step, introduces a loose, configurable coupling by means of a sequence
and steps. The responsibility for the actual boot sequence now lies at
the request handler taking care of the request.

Request handlers must now be explicitly registered in the package
class. The interface to be implemented was moved to the Core namespace.

Resolves: #32333
Releases: 1.1

Change-Id: I467a2a3b02167087a23dba1eb64d1a9747eac8ba