[TASK] Optimized package manager and configuration manager initialization
This patch introduces a somewhat lazier initialization for the
Configuration Manager's cache and the scanning of available packages
and loading of package state information.

NOTE: On the first hit, the modified Package Manager will introduce
a new format for the PackageStates.php. From now on the configuration
in that file also contains a version number which allows for automatic
migration of that file in the future.

Change-Id: I690baafd2fafa24dd104950eea82b10815280427
Related: #32333
Releases: 1.1