[FEATURE] Freezable Cache Backend
This introduces a new type of cache backends which supports freezing.
In a frozen state no modifications are allowed and thus enables the
cache implementation to optimize its tasks for speed.

The FileBackend was extended to support freezing and is now used for
the class cache in Production context. Measured on the blog example,
this change improved number of transactions per second by 29 %.

Change-Id: I1bd212157516875507ce7677515d14e8a611aa73
Resolves: #32398
Releases: 1.1