[TASK] Improved Class Loader
This patch introduces an improved loading strategy based on symbolic
links leading directly to the Classes directory of each package.
The symlinks reside in a ".Shortcuts" directory below the general
Packages folder and are created whenever the PackageStates.php file
is saved by the Package Manager.

This patch also contains minor speed improvements in the Proxy Compiler
and the Object Manager (getCaseSensitiveObjectName).

Releases: 1.1

Change-Id: I90bd40267f3fff49ab3faa6c68c2ac8feee2d070