[BUGFIX] (Utility): ArrayMergeRecursiveOverrule does not override arrays with simple types
If the first element contains an array, and the second element contains a simple
type at a certain position, the ORIGINAL array is used. That bug prevents to reset
array in configuration for example.

A full example is given in issue #32984.

I also added first unit tests for the ArrayMergeRecursiveOverrule function.

Fixes: #32984
Releases: 1.0, 1.1
Change-Id: I676efcb8c6de9e376f7dd7ea364b3c3a99b1625a