[BUGFIX] Make BCrypt the default hashing strategy (again)
In https://review.typo3.org/5756 the default hashing strategy was
changed to BCrypt. Later, in https://review.typo3.org/6598, support for
multiple strategies was added. But that change make PBKDF2 the default
again.

This change fixes that and makes the SaltedMd5 strategy available in
the YAML file as well (for completeness).

Change-Id: Icb1886a63031ae8393c391a99f7616cfb0a35b96
Fixes: #32991
Releases: 1.1