[FEATURE] Code migration tool
This adds support for "code" migrations. As with database migrations,
steps needed to adjust existing code to changes done in some package
can be delivered in a migration file with said package.

A new command flow3:core:migrate can be used to apply migrations to
all packages still in need of that migration. The tool requires git
to be available and will work only on code in clean git working copies.

Change-Id: Id237481ae651f94f2ae93d73544efb94d9d19734
Resolves: #29389
Releases: 1.1