[BUGFIX] Fixes wrong locale in XLIFF models
If a translation is requested but no (exact) match is found,
a different locale may take effect (e.g. en instead of en_UK).

The XLIFF model created will refer to the en XLIFF file, but
claim it's locale is en_UK. This leads to further problems
with translations (like missing labels if no target elements
are present in XLIFF files).

This change fixes that…

Change-Id: I2f2c3d0ba23260299a3bc4397d7d8ad9021a5b36
Fixes: #34170
Releases: 1.1