[FEATURE] Implement file based encrypted key authentication
This commit introduce a new password token, which works like the username
password token but only requires a password.

The file based simple key provider authenticates a password token using the
file based simple key service to get the matching password in the key which
is configurable for every instance of the provider. Furthermore the service
is capable of storing, removing, getting, checking and generating keys.
Optional values for generation key length and password hashing strategy is
configurable for the service.

Change-Id: I3f388737c309bc69dcffa4bc593378674ed87952
Resolves: #34648
Releases: 1.1