[FEATURE] Introduce the magic role "Anonymous"
This introduces a second magic role (besides the
role "Everybody") which is active as long as nobody
is authenticated. This allows for designing policies
in a way that anonymous users still have access to
certain resources.

Also fixes a small bug in the PolicyService which
previously did not return the Everybody or Anonymous
role on getRoles().

Change-Id: I1d5ff3bbf681ca73f3be69283491d8fd91900349
Resolves: #34900
Releases: 1.1