[!!!][BUGFIX] Changes the default for nullable columns to FALSE
This change makes columns not nullable by default, as is the case with
vanilla Doctrine 2.

This is a breaking change if you rely on NULL values to be returned
to detect special cases, like for not using unset date properties.
In those cases you need to mark your property with the Doctrine
annotation: Column(nullable=true)

Change-Id: I9537b222a0376cf87c0e547c662b90ccb1532866
Fixes: #35059
Releases: 1.1