[!!!][BUGFIX] Alternative entry points can be configured again
This fix allows for using other than the pre-defined WebRedirect
entry points (for example the HTTP Basic entry point).

In order to avoid confusion, this patch also streamlines the setting
names for authentication providers. It makes sure that only one entry
point can be defined at all and adjusts the naming of other settings
to be in line with other FLOW3 setting conventions.

The keys of settings for the authentication provider have changed as
follows:

DefaultProvider:
  provider:          '(a class name)'
  providerOptions:   [ ... options ... ]
  token:             '(a class name)'
  entryPoint:        '(a class name)'
  entryPointOptions: [ ... options ... ]

Resolves: #33495
Releases: 1.1
Change-Id: Ic8c3c7043decc9be531d80a1c016cd0f4624e4cf