[FEATURE] HTTP Client for Functional Testing
This adds a virtual HTTP client which supports pluggable request engines
for different purposes. The first scenario in which we'll use this new
"Browser" is for seamless functional testing of controllers and other
features which should be tested through the whole request / response
chain.

The FunctionalTestCase now provides a new class property called
"testableHttpEnabled" which, if set to TRUE, initializes a new browser
for each test and provides it as $this->browser.

Also adds a bunch of functional tests for the MVC ActionController using
this new feature.

Resolves: #35706
Related: #35243
Releases: 1.1
Change-Id: Idb1b9b8eb6aedf2c5540ac309c264d9b3bd26a84