[FEATURE] Track all configuration changes
This change set greatly improves the file monitoring of FLOW3's
system relevant files, such as class files, settings, object
configuration, policies and translations.

In Development context, changes to any of the above will be
tracked and the related caches are flushed. For performance
reasons, the new mechanism does not rely on the tagging
feature of the cache framework, but uses the new Simple
File Backend and rather flushes a few caches too many than
missing an important change.

Change-Id: I52b9f659740aef5d929a102ef5902b1fd1d2a588
Resolves: #5849
Resolves: #33811
Related: #35912
Releases: 1.1