[FEATURE] Allow "configurable" classes path in packages
Package instances can be asked for their classes path, which is now set
determined by the PackageManager and stored in PackageStates.php.

It can - in the future - thus be read from a Composer file and
does no longer have to be the hardcoded "Classes".

This change removes the use of the .Shortcuts folder again and makes
"overlapping" package keys possible (e.g. TYPO3.JobQueue and
TYPO3.JobQueue.Beanstalk).

Change-Id: If6f84402f03ead963b1eed557a1cf92a44a76dc5
Resolves: #36337
Releases: 1.1