[FEATURE] cURL request engine for HTTP browser
This provides a new request engine, based on cURL,
which can be used with the virtual browser contained
in FLOW3's HTTP foundation. It can be used to send
requests and receive responses via HTTP.

This patch also contains a few minor modifications in
the RequestEngineInterface and InternalRequestEngine
class.

Releases: 1.2
Resolves: #40414
Change-Id: Id58e791185c8d595de3228cd2f3874b46ab479a0