[!!!][FEATURE] Make cache backends more flexible
This change makes cache backends more flexible. It adds a new interface
"TaggableBackendInterface" and a new "SimpleFileBackend". The latter
does not support tagging but is faster in certain cases (where with a
lot of tags/entries flushing everything is faster than flushing by tag).

If you have implemented a cache backend and support tagging, you must
add the "TaggableBackendInterface" to the list of interfaces implemented
by your backend.

In addition this change makes the PDO backend a little more robust and
resolves some code style issues.

Resolves: #36563
Related: #34875
Extbase Issue: #36564
Releases: 1.1

Change-Id: I2b6c315493e983c50572026c1605ae5ad4099284