[!!!][BUGFIX] Enable settings override with empty array value
This change updates the behavior of arrayMergeRecursiveOverrule such
that empty arrays are also treated as empty values. If you used this
function before it can lead to different results after the change.

Before the change an empty array value was always merged
into the first array, even if emptyValuesOverride was set to TRUE.
With the new behavior an empty array value overrides a given input
value if emptyValuesOverride is set. This allows to override settings
with an empty array in the settings hierarchy.

This change is breaking if your settings hierarchy depends on
empty array values not overriding a parent configuration.

Resolves: #36720
Releases: master, 2.0
Change-Id: Ib7b93e62b59cea4c3d706eb39a66e01f2c44a22e