[TASK] IdentityRoutePart should return __identity array
IdentityRoutePart::matchValue() currently returns the identifier
of the ObjectPathMapping record directly.
This makes it difficult to merge arguments with data from the
requestBody (e.g. for PUT requests).

This change adjusts the route part so that it sets $this->value
to an array in the format array('__identity' => '<Identifier>')

Change-Id: I8b8cca1e20a2caf85ba51680ca79c4d74c16ed8b
Fixes: #37403
Related: #37402
Releases: 1.1, 1.2