[!!!][FEATURE] Check validator options and make them self-documenting
Validators now must declare any options they accept using the class
property $supportedOptions. It is indexed by option name and
holds an array for each option giving the default value,
a description, type and an optional required flag.

This is used to check accepted options and to generate option
documentation for the validator reference documentation.

Change-Id: I2b32130840892417214cf50cad772190fc2576c0
Resolves: #37820
Releases: 1.2