[BUGFIX] Only map file uploads to arguments if not empty
Currently the HTTP request converts the FILES superglobal
to request arguments even if no file has been uploaded.
This leads to an exception when submitting a form without
uploading a file because the property mapper receives a
corrupt source array in that case.

This adjust the HTTP request to only map values from the
FILES superglobal to arguments if a file has been uploaded.

This also tweaks the ResourceTypeConverter to accept
an empty source. Besides the converter now only outputs
filer errors (for too large or corrupt files) to the user.
Server errors (e.g. non-existing/read-only tmp directory)
are logged to the systemLogger.

Change-Id: I31519231086cd71feff73ff675806d0382ae6b40
Fixes: #38184
Releases: 1.1, 1.2