[!!!][BUGFIX] Fix return values of Doctrine\Repository magic methods
Overrides __call from EntityRepository to make the methods return
the expected QueryResultInterface. Also adds support for countBy().

This is marked as a breaking change because if you relied on the
specific return values before, you need to adjust your code.

Along the way the code inside __call() is simplified.

Change-Id: Iac08c917a07ceb679b09f254b1167401ced6c77c
Fixes: #38841
Releases: 1.1, 1.2