[!!!][TASK] Remove FLOW3's magic timezone setting behavior
FLOW3 used to magically set a default timezone if there's none set
(in order to avoid the PHP notice being thrown, resulting in an
exception).

Because every default time zone will be the wrong one for most people;
and to raise awareness of having to care about the system's time zone,
this behavior is dropped.

This change is marked breaking because in cases where no timezone
is set in PHP configuration, FLOW3 will no longer work.

Change-Id: Ife9a5b78c4cdc275df0c19235c74790334218cb3
Resolves: #37438
Releases: 1.2