[TASK] Add HTTP status code to exceptions
Currently if an exception is rendered with either Debug-
or ProductionExceptionHandler the HTTP status is set to
"500 Internal Server Error".

This change adds a property "statusCode" to the FLOW3
base Exception allowing to set a custom HTTP status code.

This also sets status codes for some of the FLOW3 exceptions

Change-Id: Id65b7d78a2851953a448893f9dcfb2c63ddb2345
Related: #27798
Releases: 1.1, 1.2