[TASK] Handle errors in subprocesses properly
To be able to properly handle missing class dependencies in ClassReflection
a ClassLoader should be added to end of the ClassLoader chain that throws
an exception if it's reached, meaning the class couldn't be loaded.

Make executeCommand throw an exception on errors, rather than just echo any
command output collected.

Change-Id: I91a1f7af5bc1456d108227806133ae55a8d61139
Resolves: #39599
Releases: 1.2