[FEATURE] Configurable CLI Subrequest environment variables
When debugging with xdebug, a major difficulty is that in development
and functional tests, there are always two requests being executed:
The main request and a CLI request inside.

Now, most IDEs do not support the (optional) feature of multi-threaded
PHP debugging, effectively stopping the sub-request and also the main
request in this case.

With this change it's possible to change xdebug configuration settings
for the CLI subrequests making it possible to debug them individually.
By default the remote port is now changed from 9000 to 9001 in
Development and Testing context. So you can debug the subrequest by
configuring your IDE to listen on port 9001.

Apart from that, this change allows you to set arbitrary environment
variables by adding entries to the
TYPO3.FLOW3.core.subRequestEnvironmentVariables setting.

Releases: 1.1, 1.2
Resolves: #39648

Change-Id: I83a60e8c730f58bc17e2df5f86e99d19bc7d553b