[FEATURE] Make WebRedirect EntryPoint less Routing dependent
The WebRedirect currently expects an option uri that points
to a relative or absolute URI to redirect when trying to access
a protected action.

This change makes it possible to specify *route values* instead
so that the Routing framework can determine the correct target
URI depending on the Routes.yaml configuration.

Change-Id: I39088e10f201b77df97a9e2cbe969a91079f379e
Resolves: #39837
Releases: 1.2