[FEATURE] Improved argument type watch in HTML debug backtrace
In the debug backtrace of method calls, this gives concrete
information about the actual argument values in case of being
boolean; and reveals the amount of array elements, if such.

An array is now presented as ``array|2|`` for example, if
it has two elements.

Boolean values are not longer only displayed as ``boolean``,
but as their concrete TRUE or FALSE value accordingly. Since
the color scheme stays the same, it can still easily be
distinguished from string values.

Change-Id: Ib8367b765fb9823e3e4c205552965caa38c9c1bc
Resolves: #39879
Releases: 1.2