[!!!][BUGFIX] Misused JoinColumns in ManyToMany
With this fix the JoinColumns are created and used in the same
way regardless if you have a JoinTable annotation.

This is (almost) a cosmetic fix as it works without, but can lead
to confusion if looking at the database.

Breaking change as the JoinColumns could change if you annotated
with JoinTable.

Change-Id: I826a89f7dccfc68cf53ec31fbe66929c02c0a70a
Fixes: #33397
Releases: 1.2