[!!!][FEATURE] Custom Error Renderers
This extends the exception handlers so that they can render arbitrary
templates based on the status code and/or class name of the thrown
exception.

The template to be used and variables that should be passed to the
template are configured in so called "renderingGroups" in the options
of the exception handler.

This is a breaking change because it deprecates the provided
NotFoundController & NotFoundView.

This patch also contains the related documentation, including general
information about FLOW3's error handling.

Also included in this patch is an update to the YAML schema definitions.

Furthermore, removed all hints on "FLOW3" from the production exception
handler.

Change-Id: I11c34dece6a880027ac199319edf5f62221a01e9
Resolves: #39423
Releases: 1.2