[TASK] Tweak install script behavior, add (example) configuration
With this change the composer post install/update script will discern
between "Essentials", being files that will always be copied - even if
they already exist in their target location - and "Defaults", being
files that will be copied only if not already existing.

This is used to add the example configuration files as essentials and a
Routes.yaml as default.

Change-Id: I836e08a3df6f3facba19eebd97d2e5ad06597f37
Related: #41031
Releases: 1.2