[BUGFIX] Fix monitoring of changed class files
This change fixes the resolving of the class name from a changed file
since the PSR-0 changes caused some problems with the assumption of
the class name layout in that place. The change keeps the old
behavior for test files since these are not yet updated to a new
layout.

Resolves: #41499
Releases: 1.2
Change-Id: Idc1d9d48998a0de0ecd07f834baa769ddf1e5ba4