[BUGFIX] Apply PSR-0 code migration only to non-PSR-0 packages
This change adds a condition to prevent the code migration to move
classes inside the Classes directory if it is already in PSR-0 layout.

Additionally an already existing composer.json will not be overwritten
and a safeguard was added to AbstractMigration that checks if a file
that should be deleted actually exists.

Change-Id: I33f7b0b126e2d15c5d057b85bb10b0d6c2b32ab0
Related: #41848
Releases: 1.2