[BUGFIX] Remove unnecessary call for code compilation
This change removes the forced compile command subrequest
in a non Production context. Proxy classes are only
initialized if the code cache is not up to date or with
a special second check after file monitoring was run.

This speeds up the development context quite enormous
if no code was changed.

Change-Id: Ifd06e26d340a501fe3a0226b70c6dda3c4035cde
Fixes: #41818
Releases: 1.2