[FEATURE] Allow integer value as timestamp in DateTimeConverter
The DateTime TypeConverter cannot transform integer values to
a DateTime object because of a strict check for string.

Accept integer values and also assume plain digits to be
a timestamp and use "U" as date format if no additional
configuration is given.

Additionally change the is_numeric() checks to ctype_digit()
as is_numeric() would also allow float values.

Change-Id: I7d9f129fa574922f62396d508a7d2579469637a5
Resolves: #41989
Releases: 1.2