[BUGFIX] Introduce HTTP Response::createFromRaw()
The Response::createFromRaw method halted if a Set-Cookie header
arrives due to lacking handling.

This brings an implementation to parse the Set-Cookie directive
as described in RFC 6265 Section 5.2. The Response::createFromRaw
method passes the Set-Cookie header data to this new implementation
in order to create a Cookie object from it.

Releases: 1.1, 1.2
Change-Id: I7c5eff63d7ae873d322aef8c3293af693fd35bf2