[FEATURE] Identify key pair by fingerprint in RsaWalletService
This change uses SSH compatible fingerprints of the RSA public
key instead of a random UUID to store keys. This allows to have
a consistent reference of keys across systems.

This change is backwards-compatible with the old behavior, because#
stored keys can still be used with their generated UUID.

Change-Id: If9652838f013e219c42c18df2765db79ebf28296
Releases: 1.2
Resolves: #42355