[BUGFIX] Fix FlowSession initialization
This patch fixes the FlowSession initialization for non HTTP-Requests.
Previously the initializeObject() method was trying to retrieve the
HTTP request even from a CLI request handler, resulting in a fatal
error.

Additionally this patch now defines a default storage backend
(FileBackend) for FlowSession. Therefore it's now sufficient to enable
FlowSession in Objects.yaml for getting started.

The garbage collection now is more reliable and triggered when a session
is resumed instead of when the shutdownObject() method is called.

Resolves: #42801
Releases: 1.2
Change-Id: I9f9d03600fd69cfd2d03b585bbcc75009e80d1e3