[FEATURE] Allow custom custom configuration files in ConfigurationManager
Using the added methods it is possible to register new "types" of
configuration. The set of processing types remains constant, but one
can now e.g. register "Assets" as configuration type and use the
"Settings" processing rule for that. This way it it possible to move
settings for assets into a dedicated file, "Assets.yaml".

The change adds

- resolveConfigurationProcessingType() method
- registerConfigurationType() method
- 2 tests to add a new ConfigurationType
- emitConfigurationManagerReady() to register new ConfigurationTypes during boot

Change-Id: I16b1de8e7c3ec826240c273a87c5e2a095815a86
Resolves: #37854
Releases: master