[FEATURE] Allow configuration processing in code migrations
This adds the new method ``processConfiguration`` to AbstractMigration.

It is given a configuration type, a processor and an optional flag
stating whether or not the processed configuration should be saved.

The processor is a Closure and is handed the parsed configuration of the
package under migration, for each found file separately. If you need to
modify the configuration, make sure to use pass-by-reference for the
processor configuration parameter in addition to setting the $saveResult
flag to TRUE.

Change-Id: I06bdc8cd1a5e3ec04ad15b058a65d182720e7044
Resolves: #43089
Releases: 1.2