[BUGFIX] Fix handling of paths with dots in getLocalizedFilename()
When a path like resource://TYPO3.Neos/Public/JavaScript/ is handed to
getLocalizedFilename(), it tries to check for the existence of
resource://TYPO3.en.Neos/Public/JavaScript/ - which breaks, unless you
happen to have a package TYPO3.en.Neos activated.

This change fixes this by using the basename in the check for an
extension.

Change-Id: I3d381f11ff9d1e27818fc0f686ddb2837c820193
Fixes: #43195
Releases: 1.2