[FEATURE] Remove trailing slash from routes for empty subroutes
Flow throws an exception if a route contains a trailing slash in
order to avoid duplicate slashes in URIs and to make it possible
to add some kind of global default suffix to all URIs.
When using sub routes and a prefix with a trailing slash in the
uriPattern of the main route (e.g. uriPattern: 'blog/<BlogSubroutes>')
this leads to an unwanted exception if there is an empty (fallback)
subroute because the merged uriPattern would contain a trailing
slash ("blog/").

This change strips trailing slashes from merged uriPatterns if the sub
routes uriPattern is empty.

Change-Id: I3499600774d62780982e707c4a7eff738f52b703
Resolves: #31777
Releases: 1.2