[BUGFIX] Check for installation of phpunit by composer
Ib4b6c68c03590444addb7c2e5af89b2a2d90051f introduced a regression
for running functional tests when composer is used in the project,
but PHPUnit is still installed using PEAR.

This change adds an additional check to verify if phpunit is actually
installed in the Libraries folder. By doing so the Bootstrap will
fall back to the PEAR installation of phpunit.

Change-Id: I7d100a9381f87b16cb1512355fee236d2b0f7ea2
Resolves: #43377
Releases: 1.2