[!!!][TASK] Use \ instead of _ for entity resources in Policies
When defining entity resources in a Policy.yaml file, the entity type
had to be the class name with \ replaced by _. With this change the
regular class name is to be used instead.

A code migration to adjust Policy.yaml files is part of the change.

Change-Id: I6e894e49b12a6c01692a3ca5c6e11eb9794b6657
Resolves: #43629
Migration: TYPO3.Flow-201212051340
Releases: 1.2