[FEATURE] ObjectConfiguration gets name from annotation
Currently, an object configuration's property's class must
explicitly configured with its name, even if the name is
implied by the annotation of the intended property.

This change falls back to that annotated class name, if
the name is not explicitly mentioned. For example, this
configuration now will work::

 'Acme\Acme\SomeClass':
   properties:
     'someProperty':
       object:
          # the type of 'someProperty' will be inferred from the var annotation now
          # previously it had to be specified via `name`
         arguments:
           1:
             value: 'SomeConstructorArgument'

…as long as the mentioned property ``someProperty`` has a
proper ``@var`` annotation revealing the class name.

Change-Id: Ic84f2605856b43aab04672da4b75b3bc36b5e196
Resolves: #44157
Releases: master