[FEATURE] Allow influencing ini usage for sub process
With a newly introduced configuration directive
``TYPO3.Flow.core.subRequestPhpIniPathAndFilename`` it
is now possible to control whether to use an ini file
or not for sub processes. The allowed values for this
setting might either be NULL or not set to use the
parent process' ini file; or might be a path to a
specific ini file.

If this setting is set to ``false``, no explicit ini file
will be used for sub processes.

Additionally, this patch refactory the actual command
generation out of the execution method in order to make
it unit-testable, which is also included in this patch.

Change-Id: I7bae89d46ef7ed017526a1a2f7c5f328e62a0c76
Resolves: #45178
Releases: master