[!!!][FEATURE] Lazy Dependency Injection
This introduces lazy dependency injection for dependencies injected
through Property Injection. Instead of loading the respective class
and instantiating the dependency right away, a proxy object is
injected which automatically replaces itself with the real dependency
once a method is called on it.

The @Inject annotation now accepts an optional argument "lazy" which
allows for turning off lazy dependency injection in specific cases.

Please read the new section in the Object Management chapter of
the Flow manual, included in this change set.

Change-Id: Ia61e40708cdb3d00210df9a5630e64aa8f89b82a
Resolves: #43819
Releases: master, 2.0