[BUGFIX] Add Flow-Doctrine cache adapter
This adds a cache adapter for Doctrine. It is used for the Doctrine ORM
metadata and query cache instead of the ArrayCache used until now.

Until now the cache implementations of Doctrine that could be configured
were not usable. This change fixes that, so that actually a cache for
metadata and query caching is used.

Change-Id: I37406667197f38f66ae778254307f09fab2b0b95
Fixes: #45264
Releases: master, 2.0