[BUGFIX] Undefined index breaks logger setup in Production
With I6da4763fb94798b1d1647b6108457996a48aa128 the logger used for the
system log was made configurable. In Production context, when cached
settings are used, this breaks after an update - and even --force with a
cache flush does not help.

This change adds a safeguard against the settings not being there.

Change-Id: I3d72e8c8580f552f45a1865bd45dae01c4413cce
Related: #40738
Related: #46053
Releases: master, 2.0