[FEATURE] Transient properties can be property mapped
This change allows properties annotated as transient to be regarded
normally for the classSchema during reflection. Those properties are
marked in the classSchema as being transient and the persistence
implementations can check for that attribute instead of existence in
the classSchema only.

Therefore it is now possible to submit transient properties in web
forms.

Change-Id: I8dcd250e5a19eec80a1e98c2acef4c134f3a5482
Resolves: #36734
Releases: master