[BUGFIX] Ignoring of inheritance type in FlowAnnotationDriver broken
Due to weird circumstances the tests for the FlowAnnotationDriver were
skipped for a while, and change Ie6e46efa858a651b2f1fe4b08d78039249504dea
broke the functionality to disable inheritance of no subclasses exist
for a class annotated with an InheritanceType annotation.

This changes fixes that.

Change-Id: I06387f1f38340c34ce3b52a4e09a3ff2081a8782
Fixes: #48231
Releases: master, 2.0