[!!!][FEATURE] HTTP components for handling requests
Currently the only way to hook into the request handling process
one has to create a custom ``RequestHandler`` (in order to intercept
the default application flow very early) or use AOP (e.g. to capture
Fluid AJAX-Widget requests).

With this change request handling in Flow is much more modular and
flexible as it allows packages to modify the default request chain by
providing or configuring "HTTP components".

Background:

This is not a breaking change strictly speaking as it doesn't affect
the public API. It's marked as such though because it quite likely
breaks Flow applications with custom request handling.

If you created a custom ``RequestHandler`` make sure that it still
works as expected and check if its still needed or if a HTTP
component could replace it. The same is true if you hooked into the
request handling via AOP.

Change-Id: Iac1bd27cd1f2869e597b696c896633f14703ec40
Resolves: #52064
Releases: master