[FEATURE] Allow use of callback function in regex search/replace
Since using the e modifier for preg_replace() is no longer recommended
(deprecated as of PHP 5.5) this is an alternative. If the $replace
argument given to searchAndReplaceRegex() in code migrations is an
anonymous function it will be passed to preg_replace_callback() and
thus allows custom replace operations.

Change-Id: I6279ae1bdb2691d8dbdc2870d55afa0f57ef0a9e
Releases: master